Data being loaded successfully
Images are being shared successfully from Google gallery and hopefully
other image apps that are capable of sharing images.  App can detect if
multiple images are passed, and sets their height dynamically so that
each is an equal height.  I will also be redoing/renaming/repurposing
the app so that it is a dynamic wallpaper app instead.  Creating an
email client outright is too much work without again passing the data
to another external Android email client, which would defeat the
purpose of our app in the first place.  Code is sloppy as heck
currently.  Desperately needs some love (and comments!).  Will work on
creating a special “editing” mode tomorrow, as well as allowing the
user to set the edited image as their device’s wallpaper.