Project rename/Added decorate button
Renamed project to something more appropriate and added in a relative
layout so that I can overlay a “decorate” button over the images that
are sent through intents.  Next I need to create the “decoration”
activity, which allows the user to place little goofy icons over the
image, and then apply as the device’s background.