Created Decorate layout
Created the basic layout for the “Decorate” activity, with a basic 9
item grid to allow the user to select up to 9 different sections of
their image to add icons to.  Created a drawer layout to contain the
icons and a button for when the user is done editing.  Also hooked up
all logic to bounce back and forth between activities.