Sticker functionality working
Stickers are now applied to the area selected by the user.  Selected
image and selected area are kept in sync dynamically.  Next I need to
ensure that old onClickListeners are being removed, and then I need to
implement a method that hides the portions of the interface, or
unfilled areas that the user hasn’t added an image to, in order to take
a screenshot, and apply this as the device background.