Screenshot working
Screenshot feature mostly working.  I made a mistake in assuming that
because the background of the activity was transparent, it would
capture the MainActivity view underneath it, but it does not.  So I
need to find a solution to that next.  There may also be a scaling
issue on some devices.  Works fine on Nexus 5 physical device, but
Genymotion emulator applies a zoomed in version for some reason.  Need
to investigate further.