Initial project setup
Created a base project to begin work on.  Project will be “Shakeit!”, a
music player that allows the user to shake left or right to play next
or previous audio tracks on the android device.  Shaking the device up
and down will choose a random song.  Covering the proximity sensor will
also mute (or maybe pause) the currently playing song.  The application
will also use notifications to allow the user to view what song is
currently being played, even if the user is outside the app.  Will
probably need to create some sort of special service to run in the
background so that the music can continue playing, even if the app is
killed.  Need to do more research in this regard.