Implement MusicService
Created a service to handle music playback for the app.  Verified that
it is functioning by playing an audio clip through it.  Basic
functionality has been built such as support for advancing to the next
track or going back to the previous track.  Need to build in the motion
controls to take advantage of this next.