Got notification pending intents working
Created three different actions to be included within our persistent
notification, which is created while music is being played through the
app.  Each one calls our MusicService, which will then perform the
requested action, and update the Notification to match.  Need to
refactor the NotificationBuilder into its own method so we can take an
argument, which will allow us to determine how to update it when the
user taps a button in the notification.