Playback functionality complete
FINALLY figured out all of the problems with the playback
functionality.  Running a service that is in charge of both music
playback, AND a notification, but also bound to an Activity is an
enormous pain.  Need to build out an interface for the user now.
Probably going to have to be pretty minimalist in order to turn project
in on time