Fixed even more bugs...sigh
Fixed bugs that occur when the user attempts to play a song immediately
after launching the app from the notification instead of from the app’s
interface.  In a rare event, the music player didn’t have a data source
applied.  Also fixed another bug where the play and pause buttons could
become “un-synced” between the notification and the app interface.
(Hopefully) fixed a random bug that happens on startup where the app
fails to bind to MusicService.  Since this is not consistently
reproducible, it’s hard to debug and fix.