Added/re-enabled features and MORE bug fixes
Added a basic track counter to provide the user with feedback as to
which song is playing out how how many total.  Implemented a basic
random coloring function that will color the LinearLayout of whichever
control the user selects within MainActivity.  Fixed several bugs
relating to the next and previous buttons not properly updating
MainActivity’s interface (in particular, the play/pause button).
Unregistered our broadcast receiver properly when the user leaves
MainActivity, since it doesn’t need to be running in the background.
Instead, we check within the onResume function to set our data to the
correct values from the MusicService.  Re-enabled the motion gestures
for previous and skip.  They seem to work pretty well, but there is no
indication to the user that they even exist as of right now.  May
implement a simple dialog on program first load to communicate this.
Even more bug fixes relating to the communication between the
MusicService and MainActivity, both from within and outside of
MainActivity.  Never ending.