Week 3 Initial Commit
Added files over from week 2 since we will (hopefully) be creating a
widget around the same app.  If not, I can build out a “generic” widget
and functionality that would serve to meet the requirements for this
week, and copy it over to a new project when I hear back from Sherry.