Widget set up
Created the layout for the widget and set it’s Provider class up as
well.  Widget appears within the launcher and can be added to a home
screen.  Verified the provider “onEnabled” and “onUpdate” functions
both fire correctly.  Next will begin to hook it up to make sure that
if is run without music service having been started (or MainActivity),
that it can stand on it’s own and support playback, as well as updating
both the service and it’s notification upon interaction.  Also, need to
ensure MainActivity can communicate to it to keep everything in sync.