Play and pause widget functionality working
Created basic code to enable the toggling of the imageView for the
play/pause button contained in the widget.  Tapping it check to make
sure that the MusicService is not null and then either plays or pauses
the MusicService’s “current” song.  Need to implement logic to start
the service in the event that MainActivity hasn’t been initialized yet.