Basic widget playback working
Playback of music within the MusicService can now be control via the
widget.  Unfortunately, it requires the MainActivity to be running and
have a valid reference to the service itself, in order to call
functions.  Trying to work in separate intents through the
MusicService’s “onStartCommand” directly from the widget is doable, but
reconnecting the MusicService to MainActivity again after starting it
from the widget would be nearly impossible without another hundred
things breaking.  Because of this, the widget will check if
MainActivity is running, and if not, require the user to start it
before it will work correctly.  Also, while the widget update
MusicService’s notification, changing music from the notification does
not update the widget yet.  Working on that now.