Configure Activity setup complete
Created and nearly completed the configuration activity for the widget.
User will be able to set custom color for the “speaker” graphic, the
button colors, and the actual widgets background tint color.
Implemented a pagination imageView to provide feedback as there are 3
views the user swipes between.  Created a viewAnimator to animate
between the 3 screens.  Now I need to grab any data the user inputs and
apply it to the widget via the WidgetProvider.  Also need to find a way
to update the onscreen “mockup” of the widget to match the user’s
choices on each screen as they make them.  Also had to manually create
2 of the animations that are for some reason missing from the SDK.
Found them at AOSP repository and manually added them.