Widget Config activity done
Widget Config activity is pretty much complete.  When done, the user
clicks a button, which dynamically updates the widget to match the one
displayed in the config activity, and places it on the home screen.
There is a bug where the intents aren’t being preserved for some
reason.  Need to fix that next, otherwise the widget is basically
useless.  Also added all images to support the multiple color
configurations that the user could possibly choose.  YAY Photoshop!