Functionality complete
App should now be functionally complete.  Entering a valid url will
display a preview of the website, and a stylized button next to the
preview to launch an intent via a Native method.  Removing a valid url
resets the field, and makes the preview invisible, as well as removing
the intent button from display.  User can save their websites to shared
preferences, which are loaded in by default the next time the app is
run.  Need to add a bit more comments before turning in.  STILL need to
rename the project to something more suiting (from its original title).
 Also need to create an app icon.